rename auth.GenerateTempAccessToken() into auth.GenerateRandomString() and move it into a separate file
